appengine/requirements.txt
# Must specify Flask version still usable with Python 2.7 here
Flask==1.1.1
google-api-python-client==1.7.11